run_vm.sh: add new argument options and small improvement
Add following options
1. -numa   : enable NUMA mode for single vmachine.
2. -start  : Start number of vCPU at a real machine.
3. -iplist : IP list which inter-connect'ed between machines
             by infiniband

Also, use $(expr `nproc` / 2) to get the correct number of physical
CPUs in real machines even though they are on the hyper-threading.
This is much easier to calculate how many CPUs we would like to attach
to each QEMU instance than take the number of real cores.

For instance, we have 2 real machine, m1 and m2, connected by infiniband
and each machine has 16 cores (HT). Each one should take 8 cores for -smp
option and vCPU must be started at 0 and 4 respectively with -local-cpu
4 option.

Be careful that this patch is only for default options of test-bed,
if any of you want to change the options in different environment.
Please use arguments.

Signed-off-by: Leesoo Ahn <lsahn@ooseel.net>